test: split cloud tests into per-source-file test files
Go convention pairs foo_test.go with foo.go. cloud_client_test.go is
split into bedrock_client_test.go / azure_client_test.go, and the
previously untested vertex_client.go gains vertex_client_test.go
(bundle validation, SA-JSON parse errors, credential cache reuse).
cloud_template_test.go merges into provider_template_test.go, which
already covers the TemplateManager it exercises.
